Create README.md
I am Zweli, currently working as a Digital Product Owner in the Health care industry. I am an IT professional by profession, looking to expand my proverbial horizons. As someone who has always been interested in the inner workings of data, I am taking this opportunity to finally learn about the analysis, the science and manupulaiton of data. 
My goal is to get to a position where I can provide data analyst consulting services to clients and organisations that are llooking to make an impact with their data.